Rename Java.Instance to Java.Entry
This type is the for the main java entry point singleton.
Instance is a misleading name, and in fact I want to introduce
a new type that is essentially a Java Object (java.lang.Object)
pointer, but this new type must be separate from java.lang.Object.
The new type will be Java.Instance.