Stub dts_test and Makefile changes: run TinkerPop.d.ts through tsc and tslint.
This is a somewhat crude mechanism to run the generated .d.ts file through
tsc and tslint to surface where the generated file is invalid. The make target
`test-package-out` runs this test. It is no yet part of 'make test', since it
fails so horribly. Once we succeed in generating a valid .d.ts file we can
make this part of the regular test.