Move tsTypeName from code-writer.ts to classes-map.ts.
I've decided it will be better to do more work in classes-map
and less work in Handlebars helpers. In particular, the transformations
previously done in code-writer tsTypeName() should be already present
in the json object that Handlebars processes.

This commit does part of that, but exposes how tsTypeName is incomplete.
The new unit test fails. The next commit will fix the failures.