Map excluded classes to 'any'. Always map java.lang.String to string.
We can exclude classes that we're unlikely to need, even if they are referenced
as arguments of methods of classes we do need, by mapping excluded class types
to the 'any' type.

While working on this I realized that we should exclude java.lang.String,
but always map it to 'string' instead of 'any'. This should match the behavior
of node-java, which automatically coerces string to java.lang.String (and vice versa)
as necessary.

A few miscellaneous changes also slipped in, such as reformating the handlebar
template file ts-templates/package.txt.