Generate java.d.ts instead of TinkerPop.d.ts (as Matt said all along).
This recipe works, as demonstrated by the Hello World scenario,
which runs a client script against the generated java.d.ts file.

We have to reference java types as e.g. Java.java.lang.String,
but that seems a small price to pay. And can perhaps be fixed in
a later story.