Move pom.xml to tinkerpop/pom.xml
This is in anticipation of making bin/ts-java be independent of tinkerpop.
I plan to refactor so that bin/ts-java is installed as a global tool
(npm install -g ts-java), and then run inside a directory that contains
a ts-java.json configuration file, which specifies the classpath and
other options needed to run.

I may also add some other java package to have another package to test with.