Configure bin/ts-java from package.json.
With this commit, (nearly) all of the important configuration options
are read from the package.json file instead of coming from the command
line or hard coded values in source files.

The output file path is still hard coded to 'java.d.ts' in the cwd.
The only relevant command line option is now --help, which still outputs
usage information, but that information is incomplete.