Split features/ts-java.feature into 'unit' and 'integration' features.
This is the first step towards using multiple .feature files.
The changes in this commit are purely organizational, with the exception
of the freeform text defining the feature at the top of nomimal.feature.

Per discussion with Matt, I'm identifying some tests as essentially
unit tests. I'd like cucumber to run these tests first (if make is run
with only a single job). Cucumber seems to run tests in lexical order
so I've named the directory 1-unittest.

I extracted the first two tests from ts-java.feature into 1-unittest/nominal.feature.

The remainder of ts-java.feature is moved into 2-integration/ts-java.feature.
This file will may be renamed and/or further subdivided in subsequent commits.

Due to this new directory organiation, it was necessary to add a `--require`
option to the cucumber command line in the Makefile for it to find the step
definitions.