Update lib/java.d.ts from reflection/java.d.ts.
This update is entrely optional but I think it is useful to
have a record of changes to this file due to changes in ts-java.
This is the result of removing a number of classes from
the 'required seed classes' in classes-map.ts, and then adding
to the seed classes and white list in reflection/package.json,
done in the previous commit.