Rework Makefile
The goal with this commit was to make a makefile that worked well
with `make -jN`, didn't do unnecessary work, and was smart enough
to do some tasks automatically that currently must be done manually,
such as running install after changing a .java file.

I succeeded in some, but not all of the goal. The big problem is that
I wasn't able to make the java packages completely independent of
each other. GNU Make's Static Pattern Rules are not flexible enough
to support hierarchical dependencies, so I had to accept flattened
dependencies.