Change makefile to fail whenever lib/java.d.ts is out of date.
It will be useful to be forced to keep lib/java.d.ts up to date
with the reflection/java.d.ts, as it is a final line defensive
programming. I thought about trying to make ts-java simply use
the reflection/java.d.ts file directly, but that's probably not
workable.