First pass generating aliases only when appropriate.
This was a relatively simple hack, but unfortunately it doesn't really solve the problem.
The changes to tsTypeName are sufficient, because we need to know the complete list of classes,
but currently tsTypeName is called while the classes are being crawled.

The solution is to refactor so that we crawl and identify all classes in one pass,
and then generate the ClassDefinition for each class in a second pass.

We'll do that refactoring next.