Use object_t for return type.
This commit isn't part of this branch's story (#_88154024), but it occurred to
me that it was the correct thing to do. Java methods that are declared to return
java.lang.Object will return different javascript types depending on the runtime
type. I already had tests in place where I declared the variable receiving the
function result to be of type object_t in order to handle this case. It was silly
to not declare the method's return type to be object_t.