Use preScanAllClasses to extend the seed classes.
This is another step in refactoring, with the goal of replacing the
first pass of loadAllClasses with the result of this preScan. In this
commit, preScan is just used to extend the set of seed classes,
and loadAllClasses is still called twice.

This works for most of the BDD test cases, but fails in tinkerpop,
due to namespace conflicts between java.lang and groovy.lang. The solution
to that is to use the full path including our outer module Java,
which will be providing in an upcoming commit.