Rename Java.Singleton->Java.NodeAPI, and use Java. scope in extends clauses.
After adding some groovy classes to tinkerpop we discovered two new methods
of name conflicts. The first is that groovy introduces a class Singleton,
which conflicted with the name we used for the Node API. In hindsight,
Singleton was a bad choice as it was too generic. The name NodeAPI is
more descriptive and should be unlikely to conflict with Java classes.

The second source of conflicts is arose because of the common submodule
'lang' used by java.lang and groovy.lang. The problem is easily resolved
by ensuring that in 'extends ...' clauses we use the fully scope class
path, including our outer module Java.