Finish refactor: preScan done in ClassesMap.initialize() and other changes.
This commit is the culmination of a larger refactoring.

Before the refactoring, ts-java starting crawling from a list of seed
classes, finding all referenced classes that were allowed by a 'white
list' filter.

Now ts-java first scans the index of all .jar files to find all class names.
It still applies the white list filter to that list, but the result could
be more classes than found previously.

Note that classes in the seedList are always included, even if they don't
match the white list.

Also note that java.lang and java.util classes are typically not seen
in the scan of .jar files. While it is possible to list 'java.lang.' and
'java.util.' in the white list, this is discouraged. It is better to explicitly
list the desired java.lang or java.util classes in the seed clases. If this
is not done, ts-java will have to make two complete passes over all classes.