Add BDD tests to show that private,protected,package scope fields not visible.
Note: I believe this is the behavior we want, but there is a chance that
we could expose these fields (using a different method of Class that returns
all fields). This would only be useful if the fields are exposed by node-java,
which I believe it does not do, but I haven't checked.