Rename seedClasses->classes, whiteList->packages.
This commit deprecates the use of `seedClases` and `whiteList` in the
tsjava section of the package.json, and introduces the new terms
`classes` and `packages` respectively.

I am intentionally leaving tinkerpop/package.json using the deprecated
terms so that there is some level of testing that the deprecation feature
is working.

We'll complete remove support for the deprecated terms once ts-tinkerpop
and v10 have been updated to use the new terms.