Add classCache and loadClassCache().
This is a refactoring that was enabled by previous changes,
and in hindsight should have been done previously.

Now that preScanAllClasses() identifies all of the possible
classes that we will need to load, we can load the Java.Class objects
for all classes into a cache before we start analysis.
While loading, we can filter out any non-public classes.

The method loadClass() should be renamed to getClass(),
but I'll do that in separate commit.