Use reference DefinitelyTyped, remove use of power-assert.
Perhaps this should have been two commits. The initial change
is to use the upstream reference repo for DefinitelyTyped
instead of the Redseal-co fork. This didn't work due to some
different in the .d.ts for power-assert, but since we don't
really use any power-assert specific features, I resolved
the problem by removing power-assert. We now instead use
the built-in assert module.